+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Drop 2.0 and 2.1 from support Ruby versions
 - Rename `Sanitize` as `FinalSanitizer`
+- Add `:strict` context for stricter sanitization
 
 ## 0.18.0